Moved ListOrDict to custom_types.py
