+++ qiskit_algorithms/eigensolvers/numpy_eigensolver.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2022, 2024.
+# (C) Copyright IBM 2022, 2025.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -27,7 +27,7 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 from qiskit_algorithms.utils.validation import validate_min
 from .eigensolver import Eigensolver, EigensolverResult
 from ..exceptions import AlgorithmError
-from ..list_or_dict import ListOrDict
+from ..custom_types import ListOrDict
 
 logger = logging.getLogger(__name__)
