+++ qiskit_algorithms/minimum_eigensolvers/numpy_minimum_eigensolver.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2022, 2023.
+# (C) Copyright IBM 2022, 2025.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -23,7 +23,7 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 
 from ..eigensolvers.numpy_eigensolver import NumPyEigensolver
 from .minimum_eigensolver import MinimumEigensolver, MinimumEigensolverResult
-from ..list_or_dict import ListOrDict
+from ..custom_types import ListOrDict
 
 logger = logging.getLogger(__name__)
