+++ qiskit_algorithms/time_evolvers/classical_methods/evolve.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2022, 2023.
+# (C) Copyright IBM 2022, 2025.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -24,7 +24,7 @@ from ..time_evolution_problem import TimeEvolutionProblem
 from ..time_evolution_result import TimeEvolutionResult
 from ...exceptions import AlgorithmError
 
-from ...list_or_dict import ListOrDict
+from ...custom_types import ListOrDict
 
 logger = logging.getLogger(__name__)
