+++ qiskit_algorithms/time_evolvers/variational/var_qte_result.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2023, 2024.
+# (C) Copyright IBM 2023, 2025.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -19,7 +19,7 @@ from qiskit.circuit import QuantumCircuit
 
 from ..time_evolution_result import TimeEvolutionResult
 
-from ...list_or_dict import ListOrDict
+from ...custom_types import ListOrDict
 
 
 class VarQTEResult(TimeEvolutionResult):
